+++ SourceryRuntime/Sources/Equality.generated.swift
@@ -108,6 +108,7 @@ extension FileParserResult {
         if self.path != rhs.path { return false }
         if self.module != rhs.module { return false }
         if self.types != rhs.types { return false }
+        if self.functions != rhs.functions { return false }
         if self.typealiases != rhs.typealiases { return false }
         if self.inlineRanges != rhs.inlineRanges { return false }
         if self.inlineIndentations != rhs.inlineIndentations { return false }
@@ -206,6 +207,7 @@ extension TemplateContext {
     /// :nodoc:
     override public func isEqual(_ object: Any?) -> Bool {
         guard let rhs = object as? TemplateContext else { return false }
+        if self.functions != rhs.functions { return false }
         if self.types != rhs.types { return false }
         if self.argument != rhs.argument { return false }
         return true
@@ -281,6 +283,7 @@ extension Types {
     override public func isEqual(_ object: Any?) -> Bool {
         guard let rhs = object as? Types else { return false }
         if self.types != rhs.types { return false }
+        if self.typealiases != rhs.typealiases { return false }
         return true
     }
 }